requirements.txt
flet==0.21.2
pystray>=0.19.5
Pillow>=10.0.0
psutil>=5.9.0
wmi>=1.5.1
pywin32>=306
pyinstaller>=6.0.0
